pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
]
